Implement comic fetching functionality and update navigation links
