+++ app/index.jsx
@@ -42,6 +42,9 @@ export default function IndexPage() {
       <Link href="/empty_page" style={styles.link}>
         Go to Empty Page (This is link)
       </Link>
+      <Link href="/Api-pull" style={styles.link}>
+        Go to Api Pull Page (This is link)
+      </Link>
     </View>
   );
 }
